requirements.txt
aiofiles==24.1.0
temporalio==1.6.0
PyYAML==6.0.2